refactor(frontend): Root layout route (#5275)
