+++ frontend/__tests__/routes/_oh.test.tsx
@@ -3,7 +3,7 @@ import { createRemixStub } from "@remix-run/testing";
 import { screen, waitFor, within } from "@testing-library/react";
 import { renderWithProviders } from "test-utils";
 import userEvent from "@testing-library/user-event";
-import MainApp from "#/routes/_oh";
+import MainApp from "#/routes/_oh/route";
 import * as CaptureConsent from "#/utils/handle-capture-consent";
 import i18n from "#/i18n";
